Allow specifying compiler for buiding libear
In intercept mode, libear is currently built using compiler specifed by
`--use-cc`. If that compiler is a cross-compiler then it may not able
to build libear, or the result libbear is not able to intercept build
commands because it doesn't target the build machine.

This change adds `--intercept-lib-cc` options to allow user to specify
correct compiler to build libear. To maintain backward compatibility,
its default value is value of `--use-cc`.